Fold PullProgressCallback back into DockerRestClient as an anonymous class
Reverts the extraction to a standalone class, restoring the original
inline anonymous-callback style while keeping the NPE-safe
getOrDefault(...) lookups, image-level status logging, and
"Download complete" handling. Removes PullProgressCallbackTest since
the callback is no longer independently testable as a public type.
